changed unsigned int -> size_t to fix segfault on 64-bit machine
git-svn-id: svn+ssh://login2/data/avl11/svnroot/MOOS/trunk@2363 5f621d73-a7ea-4b79-bc62-7e07601040e3

+++ Tools/Graphical/uIndexedPlayBack/MOOSPlayBackIndex.cpp
@@ -242,7 +242,9 @@ bool CMOOSPlayBackIndex::MessageFromLine(const std::string & sLine,
     m_ALog.GetNextToken(sLine, n, sKey);
     m_ALog.GetNextToken(sLine, n, sSrc);
 
-    unsigned int nData = sLine.find_first_not_of(" \t", n);
+    // this needs to be size_t and _not_ unsigned int to avoid segfaults on
+    // 64-bit machines: size_t is 64-bit, unsigned int is 32-bit
+    size_t nData = sLine.find_first_not_of(" \t", n);
     if (nData == string::npos)
         return false;
